Tested the RandomWord(int) function in Word class
Initially to test the function properly, I had to first create
a new final variables to remove the magical numbers. Then, I
created getters for the word arrays. Finally, in the WordTest
class it can be seen that the function has been tested on all
of the categories in addition to one wrong category as well.