showWord method in GameState has been refactored into different methods.
In order to make the code re-usable, the generation of a string that
shows the word and the printing out process has been split up. So now
createShowWordString is responsible for generating the string and showWord
prints it out to the player. Tests have been added to test for the
functionality.