Method added that handles user input when it is of type word
A handler has been created to check the input word and
compare it with the target word as well as doing the
appropriate operations in case they match or they don't.
The addition of all these new methods are to be able to
properly refactor the guessLetter method as currently it
is too big to be readable or tested.