Handler created for the case of input letter
The handleLetterInput will check if the guess taken was correct or not
and return a string message to the player stating so. Now most of the
helper methods are implemented and there are tests to ensure that they
are working properly. Next step would be to refactor the guessLetter
method itself.