GuessLetter method has been refactored to guess method.
The reason I did the latter was to make the old method which contained
a lot of functionality all in one place easier to read. So separating
the methods into smaller methods like the handleGuess which in turn is
split into smaller methods. In addition to refactoring guessLetter I also
fixed a problem with the handleHintInput methods which was had a bug
which I found out after testing the new guess method thoroughly.